moved location object: utility/object/Location -> quest/data/LocationData
far better for organisation
